refactor(menu): extract MenuItem to file
